+++ base/src/main/kotlin/com/commonsense/android/kotlin/base/extensions/generic/GenericExtensions.kt
@@ -1,0 +1,5 @@
+package com.commonsense.android.kotlin.base.extensions.generic
+
+object GenericExtensions{
+
+}